ページの切り分け、ConsumerWidgetに統一
