+++ lib/views/login.dart
@@ -1,0 +1,144 @@
+import 'package:firebase_core/firebase_core.dart';
+import 'package:flutter/material.dart';
+import 'package:firebase_auth/firebase_auth.dart';
+import 'package:cloud_firestore/cloud_firestore.dart';
+import 'package:flutter_riverpod/flutter_riverpod.dart';
+import 'package:table_calendar/table_calendar.dart';
+import 'package:three_line_diary/views/calendar.dart';
+
+// ユーザー情報の受け渡しを行うためのProvider
+final userProvider = StateProvider((ref) {
+  return FirebaseAuth.instance.currentUser;
+});
+
+// エラー情報の受け渡しを行うためのProvider
+// ※ autoDisposeを付けることで自動的に値をリセットできます
+final infoTextProvider = StateProvider.autoDispose((ref) {
+  return '';
+});
+
+// メールアドレスの受け渡しを行うためのProvider
+// ※ autoDisposeを付けることで自動的に値をリセットできます
+final emailProvider = StateProvider.autoDispose((ref) {
+  return '';
+});
+
+// パスワードの受け渡しを行うためのProvider
+// ※ autoDisposeを付けることで自動的に値をリセットできます
+final passwordProvider = StateProvider.autoDispose((ref) {
+  return '';
+});
+
+// メッセージの受け渡しを行うためのProvider
+// ※ autoDisposeを付けることで自動的に値をリセットできます
+final messageTextProvider = StateProvider.autoDispose((ref) {
+  return '';
+});
+
+class LoginPage extends ConsumerWidget {
+  @override
+  Widget build(BuildContext context, WidgetRef ref) {
+    // Providerから値を受け取る
+    final infoText = ref.watch(infoTextProvider);
+    final email = ref.watch(emailProvider);
+    final password = ref.watch(passwordProvider);
+    return Scaffold(
+      appBar: AppBar(
+        backgroundColor: Theme.of(context).colorScheme.inversePrimary,
+        title: Text('ログイン'),
+      ),
+      body: Center(
+        child: Container(
+          padding: EdgeInsets.all(24),
+          child: Column(
+            mainAxisAlignment: MainAxisAlignment.center,
+            children: <Widget>[
+              // メールアドレス入力
+              TextFormField(
+                decoration: InputDecoration(labelText: 'メールアドレス'),
+                onChanged: (String value) {
+                  // Providerから値を更新
+                  ref.read(emailProvider.state).state = value;
+                },
+              ),
+              // パスワード入力
+              TextFormField(
+                decoration: InputDecoration(labelText: 'パスワード'),
+                obscureText: true,
+                onChanged: (String value) {
+                  // Providerから値を更新
+                  ref.read(passwordProvider.state).state = value;
+                },
+              ),
+              Container(
+                padding: EdgeInsets.all(8),
+                // メッセージ表示
+                child: Text(infoText),
+              ),
+              Container(
+                width: double.infinity,
+                // ユーザー登録ボタン
+                child: ElevatedButton(
+                  child: Text('ユーザー登録'),
+                  onPressed: () async {
+                    try {
+                      // メール/パスワードでユーザー登録
+                      final FirebaseAuth auth = FirebaseAuth.instance;
+                      final result = await auth.createUserWithEmailAndPassword(
+                          email: email, password: password);
+                      // ユーザー情報を更新
+                      // userState.setUser(result.user!);
+                      ref.read(userProvider.state).state = result.user;
+                      // ユーザー登録に成功した場合
+                      // チャット画面に遷移＋ログイン画面を破棄
+                      await Navigator.of(context).pushReplacement(
+                        MaterialPageRoute(builder: (context) {
+                          return Calendar();
+                        }),
+                      );
+                    } catch (e) {
+                      // ユーザー登録に失敗した場合
+                      // Providerから値を更新
+                      ref.read(infoTextProvider.state).state =
+                          "登録に失敗しました：${e.toString()}";
+                    }
+                  },
+                ),
+              ),
+              const SizedBox(height: 0),
+              Container(
+                  width: double.infinity,
+                  child: OutlinedButton(
+                      child: Text('ログイン'),
+                      onPressed: () async {
+                        try {
+                          // メール/パスワードでログイン
+                          final FirebaseAuth auth = FirebaseAuth.instance;
+                          final result = await auth.signInWithEmailAndPassword(
+                              email: email, password: password);
+
+                          String userId = result.user?.uid ?? ''; // ユーザーIDの取得
+
+                          // ユーザー情報を更新
+                          // userState.setUser(result.user!);
+                          // ログインに成功した場合
+                          // チャット画面に遷移＋ログイン画面を破棄
+                          await Navigator.of(context).pushReplacement(
+                            MaterialPageRoute(builder: (context) {
+                              return Calendar();
+                            }),
+                          );
+                        } catch (e) {
+                          // ログインに失敗した場合
+                          ref.read(infoTextProvider.state).state =
+                              "ログインに失敗しました。";
+                          print(e);
+                        }
+                      }))
+            ],
+          ),
+        ),
+      ),
+    );
+  }
+}
